Unifies index movement w/ HistoryCycle. See msg.
This one's interesting.

There have been two functions to cycle through history,
- HistoryMoveIndex, which in practice was used to go up/down one entry at a time,
- HistorySeekPartial, which was meant to move to the next (or prev) string that
  starts with the current `h->seek_str`

With the latter, a user could type `cd` and cycle through the history entries
that started with `cd`, for example.

I'd been debating the usefulness of HistoryMoveIndex since I created it, and
this was the final nail in the coffin. By total accident, it turns out my
StrStartsWith function returns TRUE if the substring is an empty string, which
had the effect that HistorySeekPartial worked just fine to go up/down one
element.

It also turns out that I hadn't dereferenced `h->seek_str` in an if statement,
so HistoryMoveIndex wasn't being called at all. It was all just working fine
with the HistorySeekPartial function, save for a small issue.

Since HistorySeekPartial wasn't built for this, it didn't handle moving back to
the neutral -1 position. It also had a bug in the range, as seeking down checked
for `> 0` and not `>= 0`.

With the issues fixed and functions unified, HistoryMoveindex was deleted and
HistorySeekPartial was renamed to HistoryCycle.